requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
psutil==5.9.6
schedule==1.2.0 